backend/video_analyzer/requirements.txt
fastapi
uvicorn
yt-dlp
scenedetect[opencv]
openai-whisper
easyocr
chromadb
sentence-transformers